Added parameter "findCPin2D" to getNormalVector to be able to turn of this particular functionality.
This was in turn used to be able to calculate the r-Field on all
points and print it to file using Paraview.

To this end, added central and backward differential quotients to
the y-direction (which was previously unnecessary as the normal
Vector was only ever calculated close to the substrate boundary).

To visualize, load the Phi.xmf and on the left, under
"Point Arrays", choose sourceTermField. If you want to plot the
levelset field as well, you need to load another instance of the
file into Paraview. Just changing the "Point Arrays" parameter
does not work, for some reason.